requirements.txt
python-telegram-bot~=12.8
ujson~=3.2.0
schedule~=0.6.0
pandas~=1.1.2
numpy~=1.19.2
requests~=2.24.0